+++ tasty-plutus/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
-## Unreleased
+## 4.1 -- 2021-11-18
 
 ### Changed
+
+* Added `Value` parameter to minting policies. Added field to `MintingTest`, added parameter to `GenForMinting`, `fromArbitraryMinting`.
 
 * Rename `paysSelf` and `paysOther` into `paysToSelf` and `paysToOther` for
   consistency.